Rename VersionCheck.hpp to RegistryCheck.hpp
Done to prepare the header for more registry-related functions.
