+++ PackagingGE/Image3dAPI.nuspec
@@ -24,8 +24,8 @@
         <file src="Image3dAPI.net.targets" target="build/net/Image3dAPI.targets" />
  
         <!-- support headers -->
-        <file src="../Image3dAPI/ComSupport.hpp"   target="build/native/include/Image3dAPI/" />
-        <file src="../Image3dAPI/VersionCheck.hpp" target="build/native/include/Image3dAPI/" />
+        <file src="../Image3dAPI/ComSupport.hpp"    target="build/native/include/Image3dAPI/" />
+        <file src="../Image3dAPI/RegistryCheck.hpp" target="build/native/include/Image3dAPI/" />
         
         <!-- interface definitions -->
         <file src="../Image3dAPI/IImage3d.idl"   target="build/native/include/Image3dAPI/" />
